components/mlops/api/common/aicloudlibs/requirements.txt
# ===============================================================================================================#
# Copyright 2024 Infosys Ltd.                                                                                    #
# Use of this source code is governed by Apache License Version 2.0 that can be found in the LICENSE file or at  #
# http://www.apache.org/licenses/                                                                                #
# ===============================================================================================================#

FastAPI-SQLAlchemy
pydantic
uvicorn
databases==0.7.0 
motor
pymongo==4.3.3
kubernetes==10.0.0
regex